content: Add blog post Ach Linux tracking dot files
Add a new blog post about tracking home dot files in a fresh Arch Linux
installation.

The post is made from my notes about installing Arch Linux on a new
laptop.

This commit also adjust the 2 first posts about the Arch Linux
installation to link to this new one because the 3 of them belong the
same series of posts.